debug

+++ hw2/csf_assign02/c_drawing_funcs.c
@@ -282,13 +282,13 @@ void set_pixel(struct Image *img, uint32_t index, uint32_t color){
 }
 
 //return the square of an int64_t
-int64_t square(int64_t x){
+int32_t square(int32_t x){
     return (x*x);
 }
 
 //return the square distance of two points
-int64_t square_dist(int64_t x1, int64_t y1, int64_t x2, int64_t y2){
-    int64_t x_dist = x1 - x2;
-    int64_t y_dist = y1 - y2;
+int32_t square_dist(int32_t x1, int32_t y1, int32_t x2, int32_t y2){
+    int32_t x_dist = x1 - x2;
+    int32_t y_dist = y1 - y2;
     return (square(x_dist) + square(y_dist));
 }
